changed user._id to .id for all files
