+++ controllers/authController.ts
@@ -41,7 +41,7 @@ export const signin = async (req: Request, res: Response) => {
   const {email, password} = req.body
   try {
     const user = await User.signIn(email, password)
-    const token = createToken(user._id)
+    const token = createToken(user.id)
     res.cookie("jwt", token, { maxAge: 2 * 24 * 60 * 60 * 1000, httpOnly: true, secure: false, sameSite: "none" }) // 2 days // htppOnly - reminder
     res.status(200).json({ message: "User Signed In" })
   } catch (err: unknown) {
@@ -121,6 +121,39 @@ export const resetPassword = async (req: Request, res: Response) => {
   } catch (error) {
     console.error('Error resetting password:', error);
     res.status(500).json({ error: 'Internal server error' });
+  }
+};
+
+export const fetchUser = async (req: Request, res: Response) => {
+  const userId = req.user.id;
+  try {
+    const user = await User.findOne({
+      id: userId,
+    });
+    if (!user) {
+      res.status(404).json({ error: "User not found" });
+      return;
+    }
+
+    // Create a sanitized user object without password and API key
+    const sanitizedUser = {
+      id: user.id,
+      firstName: user.firstName,
+      lastName: user.lastName,
+      email: user.email,
+      country: user.country,
+      bankName: user.bankName,
+      bankAccountNumber: user.bankAccountNumber,
+      accountName: user.accountName,
+      apiUsage: user.apiUsage,
+      transactions: user.transactions,
+    };
+
+    res.status(200).json({ user: sanitizedUser });
+    return;
+  } catch (error) {
+    res.status(500).json({ error: error });
+    return;
   }
 };
