+++ routes/settings.ts
@@ -1,11 +1,9 @@
 import express from 'express';
 import { protectRoute } from '../middlewares/auth';
-import { fetchUser, resetAPIKey, updateUser, updateUserPayout } from '../controllers/settingsController';
+import {resetAPIKey, updateUser, updateUserPayout } from '../controllers/settingsController';
 
 const settingsRouter = express();
 
-//Fetch user details
-settingsRouter.get('/', protectRoute, fetchUser)
 
 //Update user payout details
 settingsRouter.post('/update-payout', protectRoute, updateUserPayout)
